# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Skolr school super app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: school_id
          in: query
          schema:
            type: string
        - name: class_id
          in: query
          schema:
            type: string
        - name: parent_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

  # ── CLASSES ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, profile_id, class_id, school_id, grade, created_at]
      properties:
        id:                  { type: string }
        profile_id:          { type: string }
        class_id:            { type: ["string", "null"] }
        school_id:           { type: string }
        grade:               { type: string }
        date_of_birth:       { type: ["string", "null"] }
        student_number:      { type: ["string", "null"] }
        full_name:           { type: string }
        avatar_url:          { type: ["string", "null"] }
        created_at:          { type: string }
    StudentInput:
      type: object
      required: [profile_id, school_id, grade]
      properties:
        profile_id:      { type: string }
        class_id:        { type: string }
        school_id:       { type: string }
        grade:           { type: string }
        date_of_birth:   { type: string }
        student_number:  { type: string }
    StudentUpdate:
      type: object
      properties:
        class_id:      { type: string }
        grade:         { type: string }
        date_of_birth: { type: string }
...